+++ CHANGELOG.md
@@ -20,7 +20,7 @@ Client mode:
     * Setting a bind variable to null would erroneously remove the relevant argument in the query
     * Default value in GraphQL request parameters was not managed
 * Correction of a NullPointerException when a query returns an interface containing fields with parameters
-
+* Issue #238: if a webClient overrides the default one, it may use its own property (the application property _graphql.endpoint.url_ is then no more mandatory). This allows such application to use any property of their own to define the GraphQL url
 
 ## 3.0.1